implement an example test case
VCL to test is in `example/example.vcl`. The test case is
in `example/tests/example.vtc`.

The test case sets up a backend server that receives an HTTP request
and sends an HTTP response with the body "Welcome to Lustrous!".

The test case sets up a client that sends an HTTP request and performs
assertions against the returned HTTP response.

The test case sets up a Varnish between the client and the server and
includes the logic from `example/example.vcl` that adds an "X-Served-By"
HTTP header to the response the value of which is "Lustrous".

It implies the assertions in the test case are:

* HTTP status code is 200; the transaction went fine.
* The body of the HTTP response is "Welcome to Lustrous!", originating
  from the backend server.
* The value of the "X-Served-By" header is "Lustrous", added
  by the tested VCL.

The VCL is included by its absolute path. I need to find a way to use
relative paths to make tests portable and to reduce repetitions.